Merge to test

+++ NMP-Database/NMP-Database/DBO/Tables/OrganicManures.sql
@@ -31,7 +31,7 @@
     [WindspeedID]              INT             NULL,
     [RainfallWithinSixHoursID] INT             NULL,
     [MoistureID]               INT             NULL,
-    [AutumnCropNitrogenUptake] DECIMAL (5, 2)  CONSTRAINT [DF_OrganicManures_AutumnCropNitrogenUptake] DEFAULT ((0)) NOT NULL,
+    [AutumnCropNitrogenUptake] INT  CONSTRAINT [DF_OrganicManures_AutumnCropNitrogenUptake] DEFAULT ((0)) NOT NULL,
     [AvailableNForNMax]        DECIMAL (18, 3) NULL, 
     [CreatedOn]                DATETIME2 (7)   DEFAULT (getdate()) NULL,
     [CreatedByID]              INT             NULL,
